Indexing refactoring
This is a pretty large refactoring of the way we intake, index, and feed cargo
to the shooter. It changes the X button to a "Cargo Adjust" command which
resets the cargo to where they're over the correct sensors. It can be tapped
a couple times if needed to settle the cargo in the robot correctly.

With these changes the cargo should be indexing better within the robot, especially
when there are two of them.